refactor: split disks/cli/helpers.py into a package
- helpers/progress.py — DisksProgressBar
- helpers/managed_mount.py — managed_mount context manager
- helpers/status.py — volume status probing and display
